Move classes to corresponding usage locations
