+++ src/dishka/dependency_source/decorator.py
@@ -1,13 +1,10 @@
 from __future__ import annotations
 
-from collections.abc import Callable
 from typing import Any, TypeVar, get_args, get_origin
 
 from dishka.entities.component import Component
 from dishka.entities.key import DependencyKey
 from dishka.entities.scope import BaseScope
-from dishka.exception_base import DishkaError
-from dishka.text_rendering import get_name
 from .factory import Factory
 from .type_match import get_typevar_replacement, is_broader_or_same_type
 
@@ -89,15 +86,3 @@ class Decorator:
 
     def __get__(self, instance: Any, owner: Any) -> Decorator:
         return Decorator(self.factory.__get__(instance, owner))
-
-
-class IndependentDecoratorError(ValueError, DishkaError):
-    def __init__(self, source: Callable[..., Any] | type) -> None:
-        self.source = source
-
-    def __str__(self) -> str:
-        name = get_name(self.source, include_module=True)
-        return (
-            f"Decorator {name} does not depends on provided type.\n"
-            f"Did you mean @provide instead of @decorate?"
-        )
