+++ src/dishka/registry_builder.py
@@ -14,12 +14,12 @@ from .entities.factory_type import FactoryType
 from .entities.key import DependencyKey
 from .entities.scope import BaseScope, InvalidScopes, Scope
 from .entities.validation_settigs import ValidationSettings
-from .exception_base import InvalidGraphError
 from .exceptions import (
     AliasedFactoryNotFoundError,
     CycleDependenciesError,
     GraphMissingFactoryError,
     ImplicitOverrideDetectedError,
+    InvalidGraphError,
     NoFactoryError,
     NothingOverriddenError,
     UnknownScopeError,
